Update README.md
I was told i was part of the .01% of people who speak the way I do. It had an interesting interest in names. What can I say? Hopefully with learning we can instill peace, compassion, & caring wisdom. 

Maybe we can use AI to look inward to ourselves & the eay we treat each other!